Move sessions guids to components
