+++ src/Host/Client/Impl/Microsoft.R.Host.Client.csproj
@@ -116,7 +116,6 @@
     <Compile Include="Extensions\DirectoryExtensions.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Definitions\GuidList.cs" />
     <Compile Include="Definitions\IRSessionCallback.cs" />
     <Compile Include="Definitions\RHostStartupInfo.cs" />
     <Compile Include="Extensions\RStringExtensions.cs" />
