+++ src/R/Components/Impl/InteractiveWorkflow/SessionGuids.cs
@@ -3,8 +3,8 @@
 
 using System;
 
-namespace Microsoft.R.Host.Client {
-    public static class GuidList {
+namespace Microsoft.R.Components.InteractiveWorkflow {
+    public static class SessionGuids {
         public static readonly Guid InteractiveWindowRSessionGuid = new Guid("77E2BCD9-BEED-47EF-B51E-2B892260ECA7");
         public static readonly Guid PackageManagerRSessionGuid = new Guid("61C93E8D-D24D-4012-82F4-093086A4FB08");
         public static readonly Guid IntellisenseRSessionGuid = new Guid("8BEF9C06-39DC-4A64-B7F3-0C68353362C9");
