+++ src/R/Components/Impl/Microsoft.R.Components.csproj
@@ -146,6 +146,7 @@
     <Compile Include="InteractiveWorkflow\Commands\SessionInformationCommand.cs" />
     <Compile Include="InteractiveWorkflow\Commands\TerminateRCommand.cs" />
     <Compile Include="InteractiveWorkflow\Commands\InterruptRCommand.cs" />
+    <Compile Include="InteractiveWorkflow\SessionGuids.cs" />
     <Compile Include="InteractiveWorkflow\Implementation\CarriageReturnProcessor.cs" />
     <Compile Include="Plots\Commands\ShowMainPlotWindowCommand.cs" />
     <Compile Include="Plots\Implementation\Commands\PlotHistoryCommand.cs" />
